+++ web-frontend/src/elements/routes/pages/search/SearchPanelMenuItems.tsx
@@ -24,7 +24,7 @@ const peakListPattern: RegExp =
   /^(\d+(\.\d+){0,1} \d+(\.\d+){0,1}( \d+(\.\d+){0,1}){0,1})(\n\d+(\.\d+){0,1} \d+(\.\d+){0,1}( \d+(\.\d+){0,1}){0,1})*$/;
 
 type InputProps = {
-  propertyFilterOptions?: ContentFilterOptions | undefined;
+  propertyFilterOptions?: ContentFilterOptions | null;
   initialStructure?: string;
   insertPlaceholder?: (
     e: KeyboardEvent<HTMLElement>,
@@ -385,7 +385,7 @@ function SearchPanelMenuItems({
                     })
                   }
                 />,
-                'Search by InChI or InChIKey of a compound (e.g. IKGXIBQEEMLURG-NVPNHPEKSA-N).' +
+                'Search by InChI or InChIKey of a compound (e.g. IKGXIBQEEMLURG-NVPNHPEKSA-N). In case of InChIKey search, the system checks for a partial match starting at the beginning (e.g. to find the first InChIKey block only).' +
                   ' ' +
                   defaultTooltipText,
               ),
@@ -401,7 +401,7 @@ function SearchPanelMenuItems({
               key: 'structure',
               style: {
                 width: '100%',
-                height: 750,
+                height: 650,
                 marginLeft: 0,
                 overflow: 'scroll',
               },
@@ -409,6 +409,8 @@ function SearchPanelMenuItems({
                 <StructuralEditor
                   initialSMILES={initialStructure}
                   insertPlaceholder={insertPlaceholder}
+                  width={'100%'}
+                  height={650}
                 />
               ),
             },