+++ web-frontend/src/types/RequestResponse.d.ts
@@ -1,0 +1,7 @@
+type RequestResponse<T> = {
+  data: T | null;
+  message: string;
+  status: 'success' | 'error';
+};
+
+export default RequestResponse;